project/requirements.txt
aiogram==3.7.0
yt-dlp==2024.5.27
python-dotenv==1.0.1
aiofiles==23.2.1